src/main/resources/scripts/requirements.txt
pdf2image
python-dotenv
requests
Pillow 
dashscope 